refactor: move getFractionDigitsOptions to number.ts
