+++ apps/main/src/dex/components/PagePool/PoolDetails/PoolParameters/index.tsx
@@ -15,12 +15,12 @@ import { Icon } from '@ui/Icon'
 import { ExternalLink } from '@ui/Link'
 import { TextEllipsis } from '@ui/TextEllipsis'
 import { Chip } from '@ui/Typography'
-import { formatDate, getFractionDigitsOptions, scanTokenPath } from '@ui/utils'
+import { formatDate, scanTokenPath } from '@ui/utils'
 import { breakpoints } from '@ui/utils/responsive'
 import { dayjs } from '@ui-kit/lib/dayjs'
 import { t } from '@ui-kit/lib/i18n'
 import { TokenIcon } from '@ui-kit/shared/ui/TokenIcon'
-import { copyToClipboard, shortenAddress, formatNumber, amount } from '@ui-kit/utils'
+import { copyToClipboard, shortenAddress, formatNumber, amount, getFractionDigitsOptions } from '@ui-kit/utils'
 import { requireBlockchainId } from '@ui-kit/utils/network'
 
 type PoolParametersProps = {
